compute hash from a special file if it exists
When the PREPARE command generates a file, it usually hashes the file
to compute the overall hash of the dependencies. But sometimes, the
content of a file is non-deterministic although it's semantics is the
same. In these situations, taking the hash of the file unnecessarily
produces different images.

As a solution, you can now generate a file call .cookHash_FILE,
where FILE is the file generated by PREPARE. If the PREPARE command
finds such a file in the same directory as FILE, it hashes the
.cookHash_FILE instead of FILE.